+++ Backlogs.Core/Utils/MediaObjects.cs
@@ -1,13 +1,7 @@
 ﻿using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
-using System;
 using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Backlogs.Utils.Core
+namespace Backlogs.Utils
 {
     public class FilmResult
     {
