--- backlog/UtilsCore/Keys.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Backlogs.Utils.Core
-{
-    public static class Keys
-    {
-        public static readonly string IMDB_KEY = "k_hrve7bkc";
-        public static readonly string TWITCH_CLIENT_ID = "p5kss9df35wfut72upg5gbpp0bwzvg";
-        public static readonly string TWITCH_CLIENT_SECRET = "gh25fs6op8euvo19r66vizqe8ytww4";
-        public static readonly string LASTFM_KEY = "94ffbb89403de22a94516302cbc0bfe2";
-        public static readonly string YOUTUBE_KEY = "AIzaSyAAmytwgkYhRsbvrWfFYthtukBy3uCXauQ";
-    }
-}
