+++ backlog/ViewModels/CompletedBacklogsViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using Backlogs.Constants;
 using Backlogs.Models;
 using Backlogs.Services;
-using Backlogs.Utils.Core;
+using Backlogs.Utils;
 using MvvmHelpers.Commands;
 using System;
 using System.Collections.ObjectModel;
