+++ backlog/Views/MainPage.xaml.cs
@@ -9,7 +9,7 @@ using Backlogs.Logging;
 using Backlogs.ViewModels;
 using Microsoft.Extensions.DependencyInjection;
 using Backlogs.Services;
-using Backlogs.Utils.Core;
+using Backlogs.Utils;
 
 // The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=402352&clcid=0x409
